+++ src/Entities/PayPalAccount.php
@@ -64,6 +64,34 @@ final class PayPalAccount extends Entity
     }
 
     /**
+     * @return BrowserData
+     */
+    public function getBrowserData()
+    {
+        return $this->getAttribute('browserData');
+    }
+
+    /**
+     * @param BrowserData $value
+     *
+     * @return $this
+     */
+    public function setBrowserData(BrowserData $value)
+    {
+        return $this->setAttribute('browserData', $value);
+    }
+
+    /**
+     * @param array $data
+     *
+     * @return BrowserData
+     */
+    public function createBrowserData(array $data)
+    {
+        return BrowserData::createFromData($data);
+    }
+
+    /**
      * @return string
      */
     public function getStatus()